feat: APIComponent logic moved to UsersContainer
